% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manejar_excel.R
\name{hojas_excel}
\alias{hojas_excel}
\title{Función que devuelva todas las hojas y dentro de las hojas las columnas en listas anidadas}
\usage{
hojas_excel(ruta)
}
\arguments{
\item{ruta}{Ruta del archivo excel}
}
\value{
Es un evector con las hojas del excel
}
\description{
Función que devuelva todas las hojas y dentro de las hojas las columnas en listas anidadas
}
\examples{
hojas_excel(NULL)
}
